Removed couple of unused functions.

+++ src/common/chemistry/reactions/SurfaceComplex.cc
@@ -22,7 +22,8 @@ namespace AmanziChemistry {
 
 namespace acu = Amanzi::AmanziChemistry::utilities;
 
-SurfaceComplex::SurfaceComplex() {
+SurfaceComplex::SurfaceComplex()
+{
   species_names_.clear();
   species_ids_.clear();
   stoichiometry_.clear();
@@ -39,24 +40,23 @@ SurfaceComplex::SurfaceComplex(const std::string& name,
                                const double free_site_stoich,
                                const double charge,
                                const double logK)
-    : name_(name),
-      identifier_(id),
-      charge_(charge),
-      species_names_(species),
-      stoichiometry_(stoichiometries),
-      species_ids_(species_ids),
-      surface_concentration_(0.),
-      free_site_name_("Unknown"),
-      free_site_stoichiometry_(free_site_stoich),
-      free_site_id_(-1),
-      h2o_stoichiometry_(h2o_stoich),
-      lnK_(acu::log_to_ln(logK)),
-      lnQK_(0.),
-      logK_(logK) {
-
+  : name_(name),
+    identifier_(id),
+    charge_(charge),
+    species_names_(species),
+    stoichiometry_(stoichiometries),
+    species_ids_(species_ids),
+    surface_concentration_(0.),
+    free_site_name_("Unknown"),
+    free_site_stoichiometry_(free_site_stoich),
+    free_site_id_(-1),
+    h2o_stoichiometry_(h2o_stoich),
+    lnK_(acu::log_to_ln(logK)),
+    lnQK_(0.),
+    logK_(logK)
+{
   logK_array_.clear();
-
-  set_ncomp(static_cast<int>(stoichiometries.size()));
+  set_ncomp(stoichiometries.size());
 }
 
 
@@ -71,29 +71,29 @@ SurfaceComplex::SurfaceComplex(const std::string& name,
                                const int free_site_id,
                                const double charge,
                                const double logK)
-    : name_(name),
-      identifier_(id),
-      charge_(charge),
-      species_names_(species),
-      stoichiometry_(stoichiometries),
-      species_ids_(species_ids),
-      surface_concentration_(0.),
-      free_site_name_(free_site_name),
-      free_site_stoichiometry_(free_site_stoich),
-      free_site_id_(free_site_id),
-      h2o_stoichiometry_(h2o_stoich),
-      lnK_(acu::log_to_ln(logK)),
-      lnQK_(0.),
-      logK_(logK) {
-
+  : name_(name),
+    identifier_(id),
+    charge_(charge),
+    species_names_(species),
+    stoichiometry_(stoichiometries),
+    species_ids_(species_ids),
+    surface_concentration_(0.0),
+    free_site_name_(free_site_name),
+    free_site_stoichiometry_(free_site_stoich),
+    free_site_id_(free_site_id),
+    h2o_stoichiometry_(h2o_stoich),
+    lnK_(acu::log_to_ln(logK)),
+    lnQK_(0.0),
+    logK_(logK)
+{
   logK_array_.clear();
-
-  set_ncomp(static_cast<int>(stoichiometries.size()));
+  set_ncomp(stoichiometries.size());
 }
 
 
 void SurfaceComplex::Update(const std::vector<Species>& primarySpecies,
-                            const SurfaceSite& surface_site) {
+                            const SurfaceSite& surface_site)
+{
   double lnQK_temp = -lnK_;
 
   // Need to consider activity of water
@@ -108,7 +108,8 @@ void SurfaceComplex::Update(const std::vector<Species>& primarySpecies,
 }
 
 
-void SurfaceComplex::AddContributionToTotal(std::vector<double> *total) {
+void SurfaceComplex::AddContributionToTotal(std::vector<double> *total)
+{
   for (int i = 0; i < ncomp_; i++) {
     (*total)[species_ids_[i]] += stoichiometry_[i] * surface_concentration();
   }
@@ -121,7 +122,8 @@ void SurfaceComplex::AddContributionToDTotal(
 }
 
 
-void SurfaceComplex::Display(const Teuchos::Ptr<VerboseObject> vo) const {
+void SurfaceComplex::Display(const Teuchos::Ptr<VerboseObject> vo) const
+{
   std::stringstream message;
   message << "    " << name() << " = ";
   message << free_site_stoichiometry_ << " " << free_site_name_ << " + ";
@@ -143,7 +145,8 @@ void SurfaceComplex::Display(const Teuchos::Ptr<VerboseObject> vo) const {
 }
 
 
-void SurfaceComplex::DisplayResultsHeader(const Teuchos::Ptr<VerboseObject> vo) const {
+void SurfaceComplex::DisplayResultsHeader(const Teuchos::Ptr<VerboseObject> vo) const
+{
   std::stringstream message;
   message << std::setw(15) << "Complex Name"
             << std::setw(15) << "Concentration"
@@ -155,7 +158,8 @@ void SurfaceComplex::DisplayResultsHeader(const Teuchos::Ptr<VerboseObject> vo) 
 }
 
 
-void SurfaceComplex::DisplayResults(const Teuchos::Ptr<VerboseObject> vo) const {
+void SurfaceComplex::DisplayResults(const Teuchos::Ptr<VerboseObject> vo) const
+{
   std::stringstream message;
   message << std::setw(15) << name()
           << std::scientific << std::setprecision(5)
